+++ ogc-example/src/main/scala/geotrellis/server/ogc/Main.scala
@@ -107,9 +107,8 @@ object Main extends CommandApp(
             simpleSources = conf
               .layers
               .values
-              .collect { case ssc@SimpleSourceConf(_, _, _, _, _, _) => ssc.models }
+              .collect { case rsc@RasterSourceConf(_, _, _, _, _, _) => rsc.toLayer }
               .toList
-              .flatten
             _ <- Stream.eval(IO(logOptState(
               conf.wms,
               ansi"%green{WMS configuration detected}, starting Web Map Service",